main changes: 	1. refactor log dir from logs to log         2. allow null replacement         3. enable insert/select materialization

+++ PivotGenerator.py
@@ -150,12 +150,14 @@ with session:
 		colNameIn = denormColInList[id1]
 		colNameOut = denormColOutList[id1]
 		for id2 in range(len(idList)):
-			rowAgg = '	,' + aggFunct + '( case '\
-	                  + 'when '\
-			          + rangeVar + ' = ' + quoteString + str(rangeList[id2]) + quoteString  \
-			          + ' then ' + colNameIn + ' else null end) (Title \'Column Info Source:' + colNameIn + ' - Row Condition : '\
-			          + str(rangeList[id2]) +'\') as ' + colNameOut + '_' \
-			          + str(id2+1).zfill(4) + '\n'
+			nullValue = udaExec.config['replaceNullValue']
+			rowAgg = '	, coalesce(' + aggFunct + '( case '\
+	           	      + 'when '\
+			       	  + rangeVar + ' = ' + quoteString + str(rangeList[id2]) + quoteString  \
+		        	  + ' then ' + colNameIn + ' else null end), ' + nullValue + ')' \
+		        	  + ' (Title \'Column Info Source:' + colNameIn + ' - Row Condition : '\
+		              + str(rangeList[id2]) +'\') as ' + colNameOut + '_' \
+		          	  + str(id2+1).zfill(4) + '\n'			
 			aggString = aggString + rowAgg
 
 	groupByList = udaExec.config['groupbyVarList'].split(',')
@@ -171,37 +173,48 @@ with session:
 	############# 
 	#concat final SQL 
 	#############	
-
-	pivotSql = 'select \n' + groupByString + aggString + 'FROM ' + DB + '.' + objectName + '\n' \
-		  + whereCondition + 'Group By ' +  groupByString + '\n'
-
+	run_id = udaExec.config['runNumber']
+	
+	pivotSql = 'select \n' + groupByString + aggString + ',\' ${appName} \' as application_name, \'${runNumber}\' as run_id FROM ' + DB + '.' + objectName + '\n' \
+                + whereCondition + 'Group By ' +  groupByString + '\n'
+	
     ############# 
 	#materialize if needed else only print the generated SQL
 	#############	
 
 	if (  udaExec.config['materializeFlag'] ==  'True'):
 	
-		if (  udaExec.config['replaceExistingTableFlag'] ==  'True'):
-			sql = 'drop table ${materializeDB}.${materializeTable}';
-			session.execute(sql,ignoreErrors=[3807])
 		
-		sql = 'create table ${materializeDB}.${materializeTable} as (\n' + pivotSql + ') with data ${materializePI}; ' 
-		session.execute(sql)
-		udaExec.checkpoint("Table created")			    
+		if (  udaExec.config['createTableFlag'] ==  'True'):
+			
+			if (  udaExec.config['replaceExistingTableFlag'] ==  'True'):
+				sql = 'drop table ${materializeDB}.${materializeTable}';
+				session.execute(sql,ignoreErrors=[3807])
+
+			sql = 'create table ${materializeDB}.${materializeTable} as (\n' + pivotSql + ') with data ${materializePI}; ' 
+			session.execute(sql)
+			udaExec.checkpoint("Table created")			    
+		
+			sql = 'comment table ${materializeDB}.${materializeTable} \'${materializeComment}\''
+			session.execute(sql)
+			udaExec.checkpoint("Table comment set")
+		
+			for id1 in range(len(denormColInList)):
+				colNameIn = denormColInList[id1]
+				colNameOut = denormColOutList[id1]
+				for id2 in range(len(idList)):
+					sql = 'COMMENT ON COLUMN ${materializeDB}.${materializeTable}.' + colNameOut + '_' \
+						  + str(id2+1).zfill(4) + ' AS  \'Column Info Source:' + aggFunct + '(' + colNameIn + ') - Row Condition : '\
+						  + str(rangeList[id2]) +'\';' + '\n'
+					session.execute(sql)
+			udaExec.checkpoint("Column comment set")
+			
+		else:
+			sql = 'insert into ${materializeDB}.${materializeTable} \n' + pivotSql  
+			session.execute(sql)
+			udaExec.checkpoint("Insert / Select")			
 		        
-		sql = 'comment table ${materializeDB}.${materializeTable} \'${materializeComment}\''
-		session.execute(sql)
-		udaExec.checkpoint("Table comment set")
 		
-		for id1 in range(len(denormColInList)):
-			colNameIn = denormColInList[id1]
-			colNameOut = denormColOutList[id1]
-			for id2 in range(len(idList)):
-				sql = 'COMMENT ON COLUMN ${materializeDB}.${materializeTable}.' + colNameOut + '_' \
-			          + str(id2+1).zfill(4) + ' AS  \'Column Info Source:' + aggFunct + '(' + colNameIn + ') - Row Condition : '\
-					  + str(rangeList[id2]) +'\';' + '\n'
-				session.execute(sql)
-		udaExec.checkpoint("Column comment set")
 
 	else:
 		print(pivotSql)
